requirements.txt
appdirs==1.4.2
beautifulsoup4==4.5.3
boto3==1.4.4
botocore==1.5.19
captcha==0.2.2
coreapi==2.3.0
coreschema==0.0.4
Django==1.10.6
django-appconf==1.0.2
django-filter==1.0.1
django-imagekit==4.0
django-pipeline==1.6.11
django-storages==1.5.2
django-widget-tweaks==1.4.1
djangorestframework==3.5.4
docutils==0.13.1
futures==3.0.5
itypes==1.1.0
Jinja2==2.9.5
jmespath==0.9.1
Markdown==2.6.8
MarkupSafe==1.0
nltk==3.2.2
numpy==1.12.1
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.0.0
psycopg2==2.7
pyparsing==2.1.10
PyPDF2==1.26.0
python-dateutil==2.6.0
requests==2.13.0
s3transfer==0.1.10
six==1.10.0
tqdm==4.11.2
uritemplate==3.0.0
